Changed code structure: merged category_selector.dart and category_selector_display.dart
