+++ lib/widgets/category_selector.dart
@@ -1,7 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:thender/main.dart';
-import 'package:thender/widgets/category_selector_display.dart';
-//import 'category_selector_display.dart';
 
 class CategorySelector extends StatefulWidget {
   @override
@@ -21,45 +19,103 @@ class _CategorySelectorState extends State<CategorySelector> {
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      height: 90.0,
-      color: Theme.of(context).primaryColor,
-      child: ListView.builder(
-        scrollDirection: Axis.horizontal,
-        itemCount: categories.length,
-        itemBuilder: (BuildContext context, int index) {
-          return GestureDetector(
-            onTap: () {
-              setState(() {
-                selectedIndex = index;
-                setScreen(selectedIndex);
-                //debugPrint(selectedIndex.toString());
-              });
-            },
-            child: Padding(
-              padding: EdgeInsets.symmetric(
-                horizontal: 15.0,
-                vertical: 30.0,
-              ),
-              child: Text(
-                categories[index],
-                style: TextStyle(
-                  color: index == selectedIndex
-                      ? homeScreenLink.textAndIconColour
-                      : homeScreenLink.otherItemsColour,
-                  fontSize: 25.0,
-                  fontWeight: FontWeight.w500,
-                  letterSpacing: 1.0,
+    return Column(
+      children: [
+        Container(
+          height: 40.0,
+          color: Theme.of(context).primaryColor,
+          child: ListView.builder(
+            scrollDirection: Axis.horizontal,
+            itemCount: categories.length,
+            itemBuilder: (BuildContext context, int index) {
+              return GestureDetector(
+                onTap: () {
+                  setState(() {
+                    selectedIndex = index;
+                  });
+                },
+                child: Padding(
+                  padding: EdgeInsets.symmetric(
+                    horizontal: 15.0,
+                  ),
+                  child: Text(
+                    categories[index],
+                    style: TextStyle(
+                      color: index == selectedIndex
+                          ? homeScreenLink.textAndIconColour
+                          : homeScreenLink.otherItemsColour,
+                      fontSize: 25.0,
+                      fontWeight: FontWeight.w500,
+                      letterSpacing: 1.0,
+                    ),
+                  ),
                 ),
+              );
+            },
+          ),
+        ),
+        Expanded(
+          child: Container(
+            padding: EdgeInsets.only(top: 3.0),
+            decoration: BoxDecoration(
+              color: Theme.of(context).accentColor,
+              borderRadius: BorderRadius.only(
+                topLeft: Radius.circular(40.0),
+                topRight: Radius.circular(40.0),
               ),
             ),
-          );
-        },
-      ),
+            child: getView(selectedIndex),
+          ),
+        ),
+      ],
     );
   }
 
-void setScreen (int index) {
-  CategorySelectorDisplay();
-}
+  Widget getView(int index) {
+    Widget setChild;
+    switch (index) {
+      case 0:
+        setChild = getPhotosView();
+        break;
+      case 1:
+        setChild = getMusicView();
+        break;
+      default:
+        debugPrint("SWITCH - NO VIEW SELECTED");
+    }
+    return setChild;
+  }
+
+  List<String> getListElements() {
+    var items = List<String>.generate(1000, (counter) => "Item $counter");
+    return items;
+  }
+
+  Widget getPhotosView() {
+    var listItems = getListElements();
+
+    var listView = ListView.builder(itemBuilder: (context, index) {
+      return ListTile(
+        leading: Icon(Icons.add),
+        title: Text(listItems[index]),
+        subtitle: Text("Random Place Holder: Photos"),
+        trailing: Icon(Icons.delete),
+      );
+    });
+    return listView;
+  }
+
+  Widget getMusicView() {
+    var listItems = getListElements();
+
+    var listView = ListView.builder(itemBuilder: (context, index) {
+      return ListTile(
+        leading: Icon(Icons.add),
+        title: Text(listItems[index]),
+        subtitle: Text("Random Place Holder: Music"),
+        trailing: Icon(Icons.delete),
+      );
+    });
+    return listView;
+  }
 }
